refactor(filesystem): isolate SD impl in its own TU so the linker tree-shakes it (#2773 item 1.2 follow-up to #2778)
Replaces the `FASTLED_USE_SDCARD` macro-gate landed in #2778 with a
**structural split** that lets the linker tree-shake SD code
AUTOMATICALLY — no user-side opt-in required, the macro is gone.

Implements the redesign requested on #2778 ("redesign this so that the
symbols can be tree shaken out"). Item 1.2 of #2773 the proper way.

## Why the macro gate wasn't the right answer

The first version of #2778 gated `<SD.h>` auto-detect behind
`FASTLED_USE_SDCARD`, requiring every SD-using sketch to add a
`#define` before `<FastLED.h>`. The reviewer pushed back that the
linker should handle this automatically. They were right — the root
cause was structural: `FileSystem::beginSd` and `make_sdcard_filesystem`
were compiled into the same translation unit (`fl.system+.cpp.o`) as
the rest of `FileSystem`, so the linker had to keep the whole TU alive
and dragged the SD/VFS/printf chain along.

## The structural fix

Move `FileSystem::beginSd(int)` and the `make_sdcard_filesystem(int)`
weak-fallback / strong-overrides into a NEW sub-aggregate:

```
src/fl/build/fl.system.sd+.cpp        # unity entry point (its own .o)
src/fl/system/sd/_build.cpp.hpp       # subdirectory aggregator
src/fl/system/sd/file_system_sd.h     # header pair
src/fl/system/sd/file_system_sd.cpp.hpp   # actual impl
```

The unity-build glob in `ci/meson/rglob.py` auto-discovers
`fl.system.sd+.cpp` and compiles it to its own `.o` archive member of
`libFastLED.a`. Per the unity-build linter's recursive-build rule
(`ci/lint_cpp/test_unity_build.py:189`), the parent
`fl/system/_build.cpp.hpp` does NOT need to include the new
`sd/_build.cpp.hpp` because the recursive entry point covers the
subdirectory independently.

## How the linker tree-shakes it now

- Sketch calls `fs.beginSd(4)` → linker has unresolved
  `FileSystem::beginSd` → scans `libFastLED.a` → finds it in
  `fl.system.sd+.cpp.o` → pulls that `.o` → SD library + vtable for
  `VFSImpl` + the `_svfprintf_r` printf chain come along. Works.
- Sketch doesn't call `beginSd` → no unresolved symbol → no archive
  pull → `fl.system.sd+.cpp.o` stays out of the link → the entire SD
  chain (libSD.a, libFS.a, `VFSImpl`, ~16 KB on ESP32-S3) is dropped.
  No macros, no opt-in, no user-visible change.

## Removes the macro from #2778

This PR removes the `FASTLED_USE_SDCARD` conditional that landed in
#2778. Migration: nothing — existing sketches that defined
`FASTLED_USE_SDCARD` keep working (the macro is now no-op but harmless),
and sketches that didn't define it but used `beginSd` (which was a
silent no-op in #2778) now work correctly again. Net effect for users
is "everything works as it did before #2778, but ~16 KB lighter for the
99% of sketches that never touch SD."

## Measured (ESP32-S3, Blink.ino, PIO + IDF 5.4 + xtensa-esp-elf-gcc 14.2.0)

| Build | firmware.bin | Δ |
|---|---:|---:|
| pre-#2778 master | 673,402 B | — |
| #2778 (macro gate, current master) | 657,680 B | −15,722 B |
| this PR (structural split) | **657,392 B** | **−16,010 B vs pre-#2778** |

Verification (from binary):

    xtensa-esp32s3-elf-nm firmware.elf | grep VFSImpl
    # empty after this PR

    xtensa-esp32s3-elf-nm firmware.elf | grep -E "FileSystem::beginSd|make_sdcard_filesystem"
    # empty — both are in the SD TU which was dropped

    grep "libFastLED.a(fl.system.sd\+.cpp.o)" firmware.map
    # no archive-pull line — the .o was never linked

When a sketch DOES call `fs.beginSd(5)`, the linker correctly pulls the
SD TU. Verified by `tests/fl/system/file_system.cpp` which calls
`fs.beginSd(5)` and expects success: passes with 1/1
(`bash test fl_system_file_system`).

## Verification

- [x] `bash test fl_system_file_system` — 1/1 pass (test calls
      `fs.beginSd(5)`, exercising the linked SD TU path).
- [x] `bash compile esp32s3 --examples Blink --platformio` — clean build.
- [x] `firmware.bin`: 657,392 B (vs 657,680 in #2778 macro-gate,
      −288 B further; vs 673,402 pre-#2778, −16,010 B total).
- [x] `nm firmware.elf | grep VFSImpl` — empty.
- [x] `bash lint --cpp` — clean (includes `EXPECTED_BUILD_FILES`
      update for the new aggregate).

Issue: https://github.com/FastLED/FastLED/issues/2773
Supersedes the macro-gate in #2778.
